Crate miden_verifier

Modules

math

Structs

ExecutionProof

A proof of correct execution of Miden VM.

Kernel

A list of exported kernel procedure hashes defining a VM kernel.

PrecompileVerifierRegistry

Registry of precompile verifiers.

ProgramInfo

A program information set consisting of its MAST root and set of kernel procedure roots used for its compilation.

StackInputs

Defines the initial state of the VM’s operand stack.

StackOutputs

Defines the final state of the VM’s operand stack at the end of program execution.

Word

A unit of data consisting of 4 field elements.

Enums

HashFunction

A hash function used during STARK proof generation.

PrecompileVerificationError

StarkVerificationError

Errors that can occur during low-level STARK proof verification.

VerificationError

Errors that can occur during proof verification.

Functions

verify

Returns the security level of the proof if the specified program was executed correctly against the specified inputs and outputs.

verify_with_precompiles

Identical to verify, with additional verification of any precompile requests made during the VM execution. The resulting aggregated precompile commitment is returned, which can be compared against the commitment computed by the VM.

Type Aliases

PrecompileTranscriptDigest

Type alias representing the finalized transcript digest.

PrecompileTranscriptState

Type alias representing the precompile transcript state (sponge capacity word).